requirements.txt
openai==0.27.2
pinecone-client==2.2.1
python-dotenv==1.0.0
langchain>='0.0.100'
faiss-cpu
tiktoken